update auth0 to handle user edge case. create util for redirecting to account.
