+++ app/utils/auth.server.ts
@@ -1,7 +1,10 @@
 import { Authenticator } from "remix-auth"
 import { Auth0ExtraParams, Auth0Profile, Auth0Strategy } from "remix-auth-auth0"
+import invariant from "tiny-invariant"
 import { getRequiredServerEnvVar } from "./environment"
 import { sessionStorage } from "./session.server"
+import { initPortalClient } from "~/models/portal/portal.server"
+import { initAdminPortal } from "~/utils/admin"
 
 // Create an instance of the authenticator, pass a generic with what your
 // strategies will return and will be stored in the session
@@ -10,6 +13,7 @@ export const authenticator = new Authenticator<{
   refreshToken: string
   extraParams: Auth0ExtraParams
   profile: Auth0Profile
+  portalUserId: string
 }>(sessionStorage)
 
 export type User = {
@@ -17,9 +21,10 @@ export type User = {
   refreshToken: string
   extraParams: Auth0ExtraParams
   profile: Auth0Profile
+  portalUserId: string
 }
 
-let auth0Strategy = new Auth0Strategy<User>(
+let auth0Strategy = new Auth0Strategy(
   {
     callbackURL: "/api/auth/auth0/callback",
     clientID: getRequiredServerEnvVar("AUTH0_CLIENT_ID"),
@@ -28,10 +33,33 @@ let auth0Strategy = new Auth0Strategy<User>(
     audience: getRequiredServerEnvVar("AUTH0_AUDIENCE"),
     scope: getRequiredServerEnvVar("AUTH0_SCOPE"),
   },
-  async ({ accessToken, refreshToken, extraParams, profile }) => {
-    // Get the user data from your DB or API using the tokens and profile
-    // return User.findOrCreate({ email: profile.emails[0].value })
-    return { accessToken, refreshToken, extraParams, profile }
+  async ({ accessToken, refreshToken, extraParams, profile }): Promise<User> => {
+    const portal = initPortalClient({ token: accessToken })
+    const getPortalUserIdResponse = await portal.getPortalUserID().catch((e) => {
+      console.log(e)
+    })
+
+    let portalUserId = getPortalUserIdResponse?.getPortalUserID
+
+    // handle edge case where user could have signed up via auth0 and yet not have an internal portalUserId
+    if (!portalUserId) {
+      const email = profile?._json?.email
+      const providerUserID = profile?.id
+
+      invariant(email, "email is not found")
+      invariant(providerUserID, "providerUserID is not found")
+
+      const portalAdmin = await initAdminPortal(portal)
+
+      const user = await portalAdmin.adminCreatePortalUser({
+        email,
+        providerUserID,
+      })
+
+      portalUserId = user.adminCreatePortalUser.portalUserID
+    }
+
+    return { accessToken, refreshToken, extraParams, profile, portalUserId }
   },
 )
