+++ app/utils/user.server.ts
@@ -2,6 +2,8 @@ import { redirect } from "@remix-run/node"
 import jwt_decode from "jwt-decode"
 import { Auth0Profile } from "remix-auth-auth0"
 import { authenticator } from "./auth.server"
+import { initPortalClient } from "~/models/portal/portal.server"
+import { PayPlanTypeV2 } from "~/models/portal/sdk"
 
 export enum Permissions {
   PayPlanTypes = "write:pay_plan_types",
@@ -86,4 +88,45 @@ export const getPoktId = (id: string) => {
 export const getUserProfile = async (request: Request) => {
   const user = await authenticator.isAuthenticated(request)
   return user?.profile
+}
+
+export const redirectToUserAccount = async (token: string) => {
+  const portal = initPortalClient({ token: token })
+  const accounts = await portal.getUserAccounts()
+  let account: { id: string | null; planType: PayPlanTypeV2 | null } = {
+    id: null,
+    planType: null,
+  }
+
+  accounts.getUserAccounts.forEach((acc) => {
+    if (acc) {
+      if (acc.planType === PayPlanTypeV2.FreetierV0) {
+        account = {
+          id: acc.id,
+          planType: acc.planType,
+        }
+      }
+      if (acc.planType === PayPlanTypeV2.PayAsYouGoV0) {
+        account = {
+          id: acc.id,
+          planType: acc.planType,
+        }
+      }
+      if (acc.planType === PayPlanTypeV2.Enterprise) {
+        account = {
+          id: acc.id,
+          planType: acc.planType,
+        }
+      }
+
+      if (account.id === null) {
+        account = {
+          id: acc.id,
+          planType: acc.planType,
+        }
+      }
+    }
+  })
+
+  return redirect(`/account/${account.id}`)
 }
